chore: amend tests and add a fixture
